refactor(imports): :art: use only @/ custom path
it was breaking some imports
