+++ .eslintrc.json
@@ -3,6 +3,17 @@
 		"next/core-web-vitals",
 		"./node_modules/ts-standard/eslintrc.json"
 	],
+	"plugins": ["testing-library"],
+	"overrides": [
+    // Only uses Testing Library lint rules in test files
+    {
+      "files": [
+        "**/__tests__/**/*.[jt]s?(x)",
+        "**/?(*.)+(spec|test).[jt]s?(x)"
+      ],
+      "extends": ["plugin:testing-library/react"]
+    }
+  ],
 	"parserOptions": {
 		"project": "./tsconfig.json"
 	},
