+++ app/users/[userName]/repositories/loading.tsx
@@ -1,4 +1,4 @@
-import { Spinner } from '@components'
+import { Spinner } from '@/components'
 
 export default function Loading () {
   return <Spinner />
