+++ package.json
@@ -13,9 +13,6 @@
   },
   "dependencies": {
     "@primer/octicons-react": "^19.1.0",
-    "@types/node": "20.2.5",
-    "@types/react": "18.2.8",
-    "@types/react-dom": "18.2.4",
     "autoprefixer": "10.4.14",
     "next": "13.4.4",
     "next-auth": "^4.22.1",
@@ -31,8 +28,13 @@
     "@testing-library/jest-dom": "^5.16.5",
     "@testing-library/react": "^14.0.0",
     "@testing-library/user-event": "^14.4.3",
+    "@types/node": "20.2.5",
+    "@types/react": "18.2.8",
+    "@types/react-dom": "18.2.4",
+    "@types/testing-library__jest-dom": "^5.14.6",
     "eslint": "8.41.0",
     "eslint-config-next": "13.4.4",
+    "eslint-plugin-testing-library": "^5.11.0",
     "husky": "^8.0.3",
     "jest": "^29.5.0",
     "jest-environment-jsdom": "^29.5.0",
