+++ tsconfig.json
@@ -20,7 +20,6 @@
 			}
 		],
 		"paths": {
-			"@*": ["./*"],
 			"@/*": ["./*"]
 
 		}
